order management system add and overall application is completed! Minor adjust need for production.

+++ src/main/java/com/khk/mgt/dao/PersonDao.java
@@ -3,6 +3,6 @@ package com.khk.mgt.dao;
 import com.khk.mgt.ds.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonDao extends JpaRepository<Person, Integer> {
+public interface PersonDao extends JpaRepository<Person, Long> {
     boolean existsByPhone(String phone);
 }
